Move all bulk data entry classes to single file
