--- BULKDATA/BulkDataEntry.py
@@ -1,5 +1,0 @@
-from abc import ABC
-
-
-class BulkDataEntry(ABC):
-    pass
