--- BULKDATA/__init__.py
@@ -1,3 +1,0 @@
-from BULKDATA import dynamic
-from BULKDATA import elements
-from BULKDATA import grid
