--- BULKDATA/dynamic/__init__.py
@@ -1,1 +1,0 @@
-from BULKDATA.dynamic import EIGRL
