--- BULKDATA/elements/__init__.py
@@ -1,3 +1,0 @@
-from BULKDATA.elements import cbar
-from BULKDATA.elements import cquad4
-from BULKDATA.elements import ctria3
