--- BULKDATA/grid/__init__.py
@@ -1,1 +1,0 @@
-from BULKDATA.grid import grid
